Rename reactive-stream-operators extensions
