+++ build-parent/pom.xml
@@ -246,22 +246,22 @@
          </dependency>
          <dependency>
             <groupId>org.jboss.shamrock</groupId>
-            <artifactId>shamrock-reactive-streams-operators-deployment</artifactId>
+            <artifactId>shamrock-smallrye-reactive-streams-operators-deployment</artifactId>
             <version>${project.version}</version>
          </dependency>
          <dependency>
             <groupId>org.jboss.shamrock</groupId>
-            <artifactId>shamrock-reactive-streams-operators-runtime</artifactId>
+            <artifactId>shamrock-smallrye-reactive-streams-operators-runtime</artifactId>
             <version>${project.version}</version>
          </dependency>
          <dependency>
             <groupId>org.jboss.shamrock</groupId>
-            <artifactId>shamrock-reactive-converters-deployment</artifactId>
+            <artifactId>shamrock-smallrye-reactive-type-converters-deployment</artifactId>
             <version>${project.version}</version>
          </dependency>
          <dependency>
             <groupId>org.jboss.shamrock</groupId>
-            <artifactId>shamrock-reactive-converters-runtime</artifactId>
+            <artifactId>shamrock-smallrye-reactive-type-converters-runtime</artifactId>
             <version>${project.version}</version>
          </dependency>
          <dependency>
