+++ extensions/reactive-streams-operators/smallrye-reactive-streams-operators/deployment/pom.xml
@@ -1,17 +1,32 @@
 <?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Copyright 2018 Red Hat, Inc.
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~     http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
     <parent>
+        <artifactId>shamrock-smallrye-reactive-streams-operators</artifactId>
         <groupId>org.jboss.shamrock</groupId>
-        <artifactId>shamrock-reactive-converters</artifactId>
         <version>1.0.0.Alpha1-SNAPSHOT</version>
     </parent>
+    <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>shamrock-reactive-converters-deployment</artifactId>
-    <name>Shamrock - Reactive Type Converters - Deployment</name>
+    <artifactId>shamrock-smallrye-reactive-streams-operators-deployment</artifactId>
+    <name>Shamrock - SmallRye Reactive Streams Operators - Deployment</name>
 
     <dependencies>
         <dependency>
@@ -20,7 +35,7 @@
         </dependency>
         <dependency>
             <groupId>org.jboss.shamrock</groupId>
-            <artifactId>shamrock-reactive-converters-runtime</artifactId>
+            <artifactId>shamrock-smallrye-reactive-streams-operators-runtime</artifactId>
         </dependency>
     </dependencies>
 
@@ -40,4 +55,5 @@
             </plugin>
         </plugins>
     </build>
+
 </project>
