+++ extensions/reactive-streams-operators/smallrye-reactive-streams-operators/deployment/src/main/java/org/jboss/shamrock/smallrye/reactivestreamoperators/SmallRyeReactiveStreamsOperatorsProcessor.java
@@ -14,23 +14,23 @@
  * limitations under the License.
  */
 
-package org.jboss.shamrock.rsops;
+package org.jboss.shamrock.smallrye.reactivestreamoperators;
 
-import io.smallrye.reactive.streams.Engine;
 import org.eclipse.microprofile.reactive.streams.operators.ReactiveStreamsFactory;
 import org.eclipse.microprofile.reactive.streams.operators.core.ReactiveStreamsFactoryImpl;
 import org.eclipse.microprofile.reactive.streams.operators.spi.ReactiveStreamsEngine;
 import org.jboss.shamrock.deployment.annotations.BuildProducer;
 import org.jboss.shamrock.deployment.annotations.BuildStep;
-import org.jboss.shamrock.deployment.Capabilities;
 import org.jboss.shamrock.deployment.builditem.FeatureBuildItem;
 import org.jboss.shamrock.deployment.builditem.substrate.ServiceProviderBuildItem;
 
-public class ReactiveStreamsOperatorsProcessor {
+import io.smallrye.reactive.streams.Engine;
+
+public class SmallRyeReactiveStreamsOperatorsProcessor {
 
     @BuildStep
     public void build(BuildProducer<ServiceProviderBuildItem> serviceProvider, BuildProducer<FeatureBuildItem> feature) {
-        feature.produce(new FeatureBuildItem(FeatureBuildItem.MP_REACTIVE_OPERATORS));
+        feature.produce(new FeatureBuildItem(FeatureBuildItem.SMALLRYE_REACTIVE_STREAMS_OPERATORS));
         serviceProvider.produce(new ServiceProviderBuildItem(ReactiveStreamsEngine.class.getName(), Engine.class.getName()));
         serviceProvider.produce(new ServiceProviderBuildItem(ReactiveStreamsFactory.class.getName(), ReactiveStreamsFactoryImpl.class.getName()));
     }
