+++ extensions/reactive-streams-operators/smallrye-reactive-type-converters/deployment/src/main/java/org/jboss/shamrock/smallrye/typeconverters/SmallRyeReactiveTypeConvertersProcessor.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.jboss.shamrock.rsops;
+package org.jboss.shamrock.smallrye.typeconverters;
 
 
 import io.smallrye.reactive.converters.ReactiveTypeConverter;
@@ -34,14 +34,14 @@ import java.util.Collection;
  *
  * Note that if none are found, nothing is added - so declaring this augmentation is quite useless in this case.
  */
-public class ReactiveStreamsOperatorsProcessor {
+public class SmallRyeReactiveTypeConvertersProcessor {
 
     private static final DotName REACTIVE_TYPE_CONVERTER = DotName.createSimple(ReactiveTypeConverter.class.getName());
 
     @BuildStep
     public void build(BuildProducer<ServiceProviderBuildItem> serviceProvider, BuildProducer<FeatureBuildItem> feature,
                       CombinedIndexBuildItem indexBuildItem) {
-        feature.produce(new FeatureBuildItem(FeatureBuildItem.REACTIVE_CONVERTERS));
+        feature.produce(new FeatureBuildItem(FeatureBuildItem.SMALLRYE_REACTIVE_TYPE_CONVERTERS));
         Collection<ClassInfo> implementors = indexBuildItem.getIndex().getAllKnownImplementors(REACTIVE_TYPE_CONVERTER);
 
         implementors.forEach(info ->
